Refactored reused code in python interface
